scripts/release/requirements.txt
google-api-python-client==1.7.11
PyGithub==1.45
